requirement.txt
attrs==18.2.0
Automat==0.7.0
boto3==1.9.57
botocore==1.12.57
certifi==2018.10.15
chardet==3.0.4
cloudpickle==0.6.1
coloredlogs==10.0
constantly==15.1.0
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dill==0.2.8.2
distro==1.3.0
docutils==0.14
en-core-web-sm==2.0.0
future==0.17.1
gevent==1.3.7
greenlet==0.4.15
humanfriendly==4.17
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jmespath==0.9.3
jsonschema==2.6.0
kiwisolver==1.0.1
klein==17.10.0
matplotlib==2.2.3
missingno==0.4.1
mkl-fft==1.0.6
mkl-random==1.0.1
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
numpy==1.15.4
packaging==18.0
pandas==0.23.4
pathlib==1.0.1
patsy==0.5.1
plac==0.9.6
preshed==2.0.1
PyHamcrest==1.9.0
pyparsing==2.3.0
PyPDF2==1.26.0
python-crfsuite==0.9.6
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
rasa-nlu==0.13.8
regex==2018.1.10
requests==2.20.1
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.9.0
simplejson==3.16.0
six==1.11.0
sklearn-crfsuite==0.3.6
spacy==2.0.18
statsmodels==0.9.0
tabula-py==1.3.1
tabulate==0.8.2
thinc==6.12.1
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
Twisted==18.9.0
typing==3.6.6
ujson==1.35
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
zope.interface==4.6.0
